requirements.txt
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
pandas>=2.1.2
openpyxl>=3.1.2
